tech(refactoring): Refactor classes and methods
