+++ src/main/java/com/linkedin/network/api/ExperienceController.java
@@ -3,7 +3,7 @@ package com.linkedin.network.api;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.linkedin.network.dto.ExperienceRequest;
-import com.linkedin.network.model.Experience;
+import com.linkedin.network.entity.Experience;
 import com.linkedin.network.response.ApiResponse;
 import com.linkedin.network.service.ExperienceService;
 import com.linkedin.network.util.Constants;
