+++ src/main/java/com/linkedin/network/entity/Experience.java
@@ -1,4 +1,4 @@
-package com.linkedin.network.model;
+package com.linkedin.network.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
