+++ src/main/java/com/linkedin/network/repository/LoginUserRepository.java
@@ -1,7 +1,7 @@
 package com.linkedin.network.repository;
 
 
-import com.linkedin.network.model.LoginUser;
+import com.linkedin.network.entity.LoginUser;
 
 import org.springframework.data.jpa.repository.JpaRepository;
