+++ src/main/java/com/linkedin/network/service/FeedService.java
@@ -1,7 +1,7 @@
 package com.linkedin.network.service;
 
 import com.linkedin.network.dto.FeedRequest;
-import com.linkedin.network.model.Feed;
+import com.linkedin.network.entity.Feed;
 
 import java.util.List;
