+++ src/main/java/com/linkedin/network/service/SkillsService.java
@@ -1,7 +1,7 @@
 package com.linkedin.network.service;
 
 import com.linkedin.network.dto.SkillsRequest;
-import com.linkedin.network.model.Skills;
+import com.linkedin.network.entity.Skills;
 
 import java.util.List;
