+++ src/main/java/com/linkedin/network/service/impl/LoginServiceImpl.java
@@ -1,6 +1,6 @@
 package com.linkedin.network.service.impl;
 
-import com.linkedin.network.model.LoginUser;
+import com.linkedin.network.entity.LoginUser;
 import com.linkedin.network.repository.LoginUserRepository;
 import com.linkedin.network.service.LoginService;
 import org.springframework.beans.factory.annotation.Autowired;
