+++ src/main/java/com/linkedin/network/service/impl/SkillsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.linkedin.network.service.impl;
 
 import com.linkedin.network.dto.SkillsRequest;
-import com.linkedin.network.model.Skills;
+import com.linkedin.network.entity.Skills;
 import com.linkedin.network.repository.SkillsRepository;
 import com.linkedin.network.service.SkillsService;
 import org.springframework.beans.factory.annotation.Autowired;
